Move vsimem_rmtree_toplevel to util + tests to new test_util.py
