+++ pyogrio/core.py
@@ -30,7 +30,6 @@ with GDALEnv():
         ogr_vsi_listtree,
         ogr_vsi_rmtree,
         ogr_vsi_unlink,
-        vsimem_rmtree_toplevel,
     )
 
     _init_gdal_data()
@@ -385,26 +384,3 @@ def vsi_unlink(path: Union[str, Path]):
         path = path.as_posix()
 
     ogr_vsi_unlink(path)
-
-
-def _vsimem_rmtree_toplevel(path: Union[str, Path]):
-    """Remove the parent directory of the file path recursively.
-
-    This is used for final cleanup of an in-memory dataset, which may have been
-    created within a directory to contain sibling files.
-
-    Additional VSI handlers may be chained to the left of /vsimem/ in path and
-    will be ignored.
-
-    Remark: function is defined here to be able to run tests on it.
-
-    Parameters
-    ----------
-    path : str or pathlib.Path
-        path to in-memory file
-
-    """
-    if isinstance(path, Path):
-        path = path.as_posix()
-
-    vsimem_rmtree_toplevel(path)
